refactor(settings): extract platform path autodetect to service
Move __get_darwin_paths, __get_linux_paths, __get_windows_paths,
_find_steam_root, and _find_mac_app_bundle from SettingsController
into a new PathAutodetectService class.

The controller now delegates to the service in
_on_locations_autodetect_button_clicked, keeping only the UI wiring
(path display, checkbox state, Snap warning).

-243 lines in controller, +219 lines in new service. No behavior change.
